Extract remaining routes, delete comments
